fastcamp/src/test/java/springcore/fastcamp/fastcamp/controller/PersonControllerTest.java
package springcore.fastcamp.fastcamp.controller;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MockMvcBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import springcore.fastcamp.fastcamp.dto.PersonDto;

import java.time.LocalDate;

import static org.junit.jupiter.api.Assertions.*;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@SpringBootTest
class PersonControllerTest {

    @Autowired
    private PersonController personController;

    private MockMvc mockMvc;

    @Autowired
    ObjectMapper objectMapper;

    @BeforeEach
    void beforeEach(){
        mockMvc = MockMvcBuilders.standaloneSetup(personController).build();
    }

    @Test
    void getPerson() throws Exception {
        mockMvc.perform(
                MockMvcRequestBuilders.get("/api/person/1"))
                .andDo(print())
                .andExpect(status().isOk());
    }

    @Test
    void postPerson() throws Exception {
        mockMvc.perform(
                MockMvcRequestBuilders.post("/api/person")
                        .contentType(MediaType.APPLICATION_JSON_UTF8)
                        .content("{\n" +
                                "  \"name\": \"name11\",\n" +
                                "  \"blood_type\": \"A\",\n" +
                                "  \"age\": 12\n" +
                                "}"))
                .andDo(print())
                .andExpect(status().isCreated());
    }

    @Test
    void modifyPerson() throws Exception {
        mockMvc.perform(
                MockMvcRequestBuilders.put("/api/person?id=1")
                        .contentType(MediaType.APPLICATION_JSON_UTF8)
                        .content("{\n" +
                                "  \"name\": \"name11\",\n" +
                                "  \"blood_type\": \"A\",\n" +
                                "  \"age\": 100\n" +
                                "}"))
                .andDo(print())
                .andExpect(status().isOk());
    }

    @Test
    @DisplayName("회원 이름 수정")
    void modifyUserName() throws Exception{
        //given
        mockMvc.perform(
                MockMvcRequestBuilders.patch("/api/person")
                        .param("name","hello")
                        .content("{\n" +
                                "  \"name\": \"name11\",\n" +
                                "  \"blood_type\": \"A\",\n" +
                                "  \"age\": 100\n" +
                                "}"))
                .andDo(print())
                .andExpect(status().isOk());
        //when

        //then
    }

    @Test
    void deletePerson() throws Exception{
        mockMvc.perform(
                MockMvcRequestBuilders.delete("/api/person/1"))
                .andDo(print())
                .andExpect(status().isOk());
    }

    @Test
    @DisplayName("json check")
    void checkJson() throws Exception{
        //given
        PersonDto dto = new PersonDto();
        dto.setName("Hello");
        dto.setBirthday(LocalDate.now());

        System.out.println("dto = " + toJsonString(dto));
        //when

        //then
    }

    private String toJsonString(PersonDto personDto) throws JsonProcessingException {
        return objectMapper.writeValueAsString(personDto);
    }

}